Refactor rich content helper functions: consolidate and replace literal merge tag functions with unified richContentText and richContentWith implementations across tests
